:bug:fix: confirm_num 타입 long으로 변경

+++ src/main/java/com/be/recheckbe/domain/receipt/entity/Receipt.java
@@ -30,7 +30,7 @@ public class Receipt extends BaseTimeEntity {
   private String cardCompany; // 카드사
 
   @Column(unique = true) // (nullable = false)
-  private int confirmNum; // 영수증 승인번호
+  private long confirmNum; // 영수증 승인번호
 
   @Column // (nullable = false)
   private String storeName; // 점포 이름
